# openapi.yaml
openapi: 3.0.0
info:
  title: TradeOn API
  version: 1.0.0
  description: 'TradeOn은 사용자 간 물물교환 기반의 중고 거래 플랫폼입니다.


    이 API 문서는 다음과 같은 기능들을 포함합니다:

    - 사용자 회원가입, 로그인 및 정보 조회

    - 교환 요청 및 거래 이력 관리

    - 사용자 리뷰 시스템

    - 가치 평가 게시판

    - 하이브리드 검색기능

    - 게시물 등록/삭제/조회

    - 채팅 및 거래 약속

    - 거래 데이터 분석 및 시각화


    JWT 인증이 필요한 API의 경우 Authorize 버튼을 통해 토큰을 입력해주세요.'
servers:
- url: https://673e-121-140-48-141.ngrok-free.app
  description: 로컬
...
paths:
  /api/v1/users/register:
    post:
      tags:
        - User
      summary: 사용자 회원가입 (id=userid)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegisterRequest'
      responses:
        '201':
          description: 회원가입 완료
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: 회원가입 완료
        '400':
          description: 필수 입력 항목 누락
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 이미 존재하는 사용자
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/users/login:
    post:
      tags:
        - User
      summary: 사용자 로그인
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginRequest'
      responses:
        '200':
          description: 로그인 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserLoginResponse'
        '401':
          description: 로그인 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/users/logout:
    post:
      tags:
        - User
      summary: 사용자 로그아웃 (ex.localStorage.removeItem('access_token') 등으로 클라이언트에서 토큰 제거해야함 )
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 로그아웃 처리 완료
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: User logged out (토큰 클라이언트에서 제거 필요)
  /api/v1/users/{userid}/profile:
    get:
      tags:
        - User
      summary: 사용자 프로필 조회 ((userid(int)입력))
      parameters:
        - name: userid
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 프로필 반환
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
        '404':
          description: 사용자 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/users/{userid}/reviews:
    get:
      tags:
        - User
      summary: 사용자 리뷰 조회 (userid(int)입력)
      parameters:
        - name: userid
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 리뷰 리스트 반환
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserReview'
  /api/v1/users/{userid}/trades:
    get:
      tags:
        - User
      summary: 사용자 거래 내역 조회 (userid(int)입력)
      parameters:
        - name: userid
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 거래 목록 반환
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserTrade'
  /api/v1/users/{userid}/points:
    patch:
      tags:
        - User
      summary: 사용자 포인트 수정 ('amount 값 1~50 제한')
      security:
        - bearerAuth: []
      parameters:
        - name: userid
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPointsUpdateRequest'
      responses:
        '200':
          description: 포인트 수정 성공
        '400':
          description: 필드 누락
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 사용자 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/users/{userid}/grade:
    get:
      tags:
        - User
      summary: 사용자 등급 조회
      parameters:
        - name: userid
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 등급 정보 반환
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserGradeResponse'
        '404':
          description: 사용자 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/users/{userid}:
    delete:
      tags:
        - User
      summary: 사용자 탈퇴
      security:
        - bearerAuth: []
      parameters:
        - name: userid
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 탈퇴 완료
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 사용자 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/search:
    get:
      tags:
        - Search
      summary: 게시글 키워드 검색
      description: |
        title 및 category 파라미터를 AND 조건으로 검색합니다.  
        두 파라미터 중 하나는 필수이며, 둘 다 없으면 400 에러를 반환합니다.
      security:
        - bearerAuth: []
      parameters:
        - name: title
          in: query
          required: false
          schema:
            type: string
          description: 검색할 게시글 제목 키워드
        - name: category
          in: query
          required: false
          schema:
            type: string
          description: 검색할 카테고리
      responses:
        '400':
          description: title 또는 category 파라미터가 필요합니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '200':
          description: 검색 결과 반환
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  count:
                    type: integer
                  posts:
                    type: array
                    items:
                      $ref: '#/components/schemas/SearchPostItem'
  /api/v1/search/advanced:
    get:
      tags:
        - Search
      summary: 제목 기반 고급 게시글 검색
      description: 입력한 제목(title)을 기반으로 하이브리드 검색을 수행하여 유사 게시글을 반환합니다.
      parameters:
        - name: title
          in: query
          required: true
          description: 검색할 게시글 제목
          schema:
            type: string
            example: 에어팟 프로 중고 팝니다
      responses:
        '200':
          description: 검색 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdvancedSearchResponse'
        '400':
          description: 잘못된 요청 (title 없음)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    UserRegisterRequest:
      type: object
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
          example: user123
        email:
          type: string
          example: user@example.com
        password:
          type: string
          example: password123
    UserLoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: user123
        password:
          type: string
          example: password123
    UserLoginResponse:
      type: object
      properties:
        access_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        user:
          type: object
          properties:
            id:
              type: integer
              example: 1
            username:
              type: string
              example: user123
            email:
              type: string
              example: user@example.com
    UserProfileResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          example: user123
        email:
          type: string
          example: user@example.com
        image:
          type: string
          example: https://example.com/profile.png
    UserReview:
      type: object
      properties:
        id:
          type: integer
        reviewer:
          type: string
        rating:
          type: integer
        comment:
          type: string
    UserTrade:
      type: object
      properties:
        trade_id:
          type: integer
        partner_id:
          type: integer
        status:
          type: string
    UserPointsUpdateRequest:
      type: object
      required:
        - amount
      properties:
        amount:
          type: integer
          example: 150
    UserGradeResponse:
      type: object
      properties:
        grade:
          type: string
          example: 초보상인
        points:
          type: integer
          example: 1250
...
    AdvancedSearchResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        count:
          type: integer
          example: 2
        posts:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
                example: 123
              title:
                type: string
                example: 에어팟 프로 중고 팝니다
              user_id:
                type: integer
                example: 1
              category:
                type: string
                example: 전자기기
              created_at:
                type: string
                format: date-time
                example: 2024-06-01
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: error
        msg:
          type: string
          example: 오류 메시지입니다.
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []